% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-raster.R
\name{as_qgis_argument.RasterLayer}
\alias{as_qgis_argument.RasterLayer}
\alias{as_qgis_argument.RasterBrick}
\alias{qgis_as_raster}
\alias{qgis_as_brick}
\alias{qgis_as_raster.qgis_outputRaster}
\alias{qgis_as_brick.qgis_outputRaster}
\alias{qgis_as_raster.qgis_outputLayer}
\alias{qgis_as_brick.qgis_outputLayer}
\alias{qgis_as_raster.qgis_result}
\alias{qgis_as_brick.qgis_result}
\title{Convert raster objects to/from QGIS inputs/outputs}
\usage{
\method{as_qgis_argument}{RasterLayer}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{RasterBrick}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

qgis_as_raster(output, ...)

qgis_as_brick(output, ...)

\method{qgis_as_raster}{qgis_outputRaster}(output, ...)

\method{qgis_as_brick}{qgis_outputRaster}(output, ...)

\method{qgis_as_raster}{qgis_outputLayer}(output, ...)

\method{qgis_as_brick}{qgis_outputLayer}(output, ...)

\method{qgis_as_raster}{qgis_result}(output, ...)

\method{qgis_as_brick}{qgis_result}(output, ...)
}
\arguments{
\item{x}{A \code{\link[raster:raster]{raster::raster()}} or \code{\link[raster:brick]{raster::brick()}}.}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}

\item{use_json_input}{TRUE if the outputs will be
serialized as JSON instead of serialized as a command-line argument.}

\item{output}{The result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} or \code{\link[=qgis_output]{qgis_output()}}.}

\item{...}{Passed to \code{\link[raster:raster]{raster::raster()}} or \code{\link[raster:brick]{raster::brick()}}.}
}
\description{
Convert raster objects to/from QGIS inputs/outputs
}
