% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-sf.R
\name{as_qgis_argument.sf}
\alias{as_qgis_argument.sf}
\title{Convert sf objects to/from QGIS inputs/outputs}
\usage{
\method{as_qgis_argument}{sf}(x, spec = qgis_argument_spec())
}
\arguments{
\item{x}{An object passed to a QGIS processing algorithm}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}
}
\description{
Convert sf objects to/from QGIS inputs/outputs
}
