% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-terra.R
\name{as_qgis_argument.SpatRaster}
\alias{as_qgis_argument.SpatRaster}
\alias{qgis_as_terra}
\alias{qgis_as_terra.qgis_outputRaster}
\alias{qgis_as_terra.qgis_outputLayer}
\alias{qgis_as_terra.qgis_result}
\title{Convert terra objects to/from QGIS inputs/outputs}
\usage{
\method{as_qgis_argument}{SpatRaster}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

qgis_as_terra(output, ...)

\method{qgis_as_terra}{qgis_outputRaster}(output, ...)

\method{qgis_as_terra}{qgis_outputLayer}(output, ...)

\method{qgis_as_terra}{qgis_result}(output, ...)
}
\arguments{
\item{x}{A \code{\link[terra:rast]{terra::rast()}}.}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}

\item{use_json_input}{TRUE if the arguments will be
serialized as JSON instead of serialized as a command-line argument.}

\item{output}{The result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} or \code{\link[=qgis_output]{qgis_output()}}.}

\item{...}{Passed to \code{\link[terra:rast]{terra::rast()}}.}
}
\description{
Convert terra objects to/from QGIS inputs/outputs
}
