% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-result.R
\name{is_qgis_result}
\alias{is_qgis_result}
\alias{qgis_result_clean}
\alias{qgis_extract_output_by_name}
\alias{qgis_extract_output_by_position}
\alias{qgis_extract_output_by_class}
\alias{qgis_error_output_does_not_exist}
\alias{qgis_result_status}
\alias{qgis_result_stdout}
\alias{qgis_result_stderr}
\alias{qgis_result_args}
\title{Access algorithm results}
\usage{
is_qgis_result(x)

qgis_result_clean(x)

qgis_extract_output_by_name(x, name = "OUTPUT", single = TRUE)

qgis_extract_output_by_position(x, which)

qgis_extract_output_by_class(x, class, single = TRUE)

qgis_error_output_does_not_exist(x, which)

qgis_result_status(x)

qgis_result_stdout(x)

qgis_result_stderr(x)

qgis_result_args(x)
}
\arguments{
\item{x}{An object returned by \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}.}

\item{which}{The name or index of an output.}

\item{default}{A default value if the output does not exist.}

\item{what}{Character vector of classes.
At least one class must be inherited by an element of \code{x} for that element
to be selected.}
}
\description{
Access algorithm results
}
