% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-algorithms.R
\name{qgis_algorithms}
\alias{qgis_algorithms}
\alias{qgis_has_algorithm}
\alias{qgis_has_provider}
\alias{qgis_providers}
\title{Get metadata about algorithms and processing providers}
\usage{
qgis_algorithms(query = FALSE, quiet = TRUE)

qgis_has_algorithm(algorithm)

qgis_has_provider(provider, query = FALSE, quiet = TRUE)

qgis_providers()
}
\arguments{
\item{query}{Use \code{TRUE} to refresh the cached value.}

\item{quiet}{Use \code{FALSE} to display more information about the command,
possibly useful for debugging.}

\item{algorithm}{A qualified algorithm name
(e.g., "native:filedownloader").}

\item{provider}{A provider identifier (e.g., "native")}
}
\description{
Returns metadata about algorithms and processing providers.
See the \href{https://docs.qgis.org/latest/en/docs/user_manual/processing_algs/qgis/index.html}{QGIS docs}
for a detailed description of the algorithms provided
'out of the box' on QGIS.
}
\examples{
if (has_qgis()) qgis_has_algorithm("native:filedownloader")
if (has_qgis()) qgis_algorithms()
if (has_qgis()) qgis_has_provider("native")
if (has_qgis()) qgis_providers()

}
