% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-arguments.R
\name{qgis_argument_spec}
\alias{qgis_argument_spec}
\title{Specify QGIS argument types}
\usage{
qgis_argument_spec(
  algorithm = NA_character_,
  name = NA_character_,
  description = NA_character_,
  qgis_type = NA_character_,
  available_values = character(0),
  acceptable_values = character(0)
)
}
\arguments{
\item{name, description, qgis_type, available_values, acceptable_values}{Column values of \code{arguments} denoting
the argument name, description, and acceptable values.}
}
\value{
A \code{\link[=list]{list()}} with an element for each input argument.
}
\description{
Specify QGIS argument types
}
\examples{
qgis_argument_spec()
if (has_qgis()) qgis_argument_spec_by_name("native:filedownloader", "URL")

}
