% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-terra.R
\name{qgis_as_terra}
\alias{qgis_as_terra}
\alias{qgis_as_terra.qgis_outputRaster}
\alias{qgis_as_terra.qgis_outputLayer}
\alias{qgis_as_terra.qgis_result}
\title{Convert a qgis_result object or one of its elements to a terra object}
\usage{
qgis_as_terra(output, ...)

\method{qgis_as_terra}{qgis_outputRaster}(output, ...)

\method{qgis_as_terra}{qgis_outputLayer}(output, ...)

\method{qgis_as_terra}{qgis_result}(output, ...)
}
\arguments{
\item{output}{A result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} or one of the
\code{\link[=qgis_extract_output]{qgis_extract_output()}} functions.}

\item{...}{Arguments passed to \code{\link[terra:rast]{terra::rast()}}.}
}
\description{
Convert a qgis_result object or one of its elements to a terra object
}
