% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-result.R
\name{qgis_clean_result}
\alias{qgis_clean_result}
\title{Clean processing results}
\usage{
qgis_clean_result(x)
}
\arguments{
\item{x}{An object returned by \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}.}
}
\description{
Deletes any temporary files that are defined in a
\code{qgis_result} object.
These may comprise both input and output files.
}
\seealso{
Other topics about accessing or managing processing results: 
\code{\link{qgis_as_raster}()},
\code{\link{qgis_as_terra}()},
\code{\link{qgis_extract_output}()},
\code{\link{qgis_result_status}()},
\code{\link{st_as_sf}},
\code{\link{st_as_stars}}
}
\concept{main functions to access or manage processing results}
\concept{topics about accessing or managing processing results}
