% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-cache.R
\name{qgis_delete_old_cachefiles}
\alias{qgis_delete_old_cachefiles}
\title{Delete old cache files}
\usage{
qgis_delete_old_cachefiles(
  type = "all",
  age_days = NULL,
  quiet = FALSE,
  startup = FALSE
)
}
\arguments{
\item{type}{A string; either \code{"all"}, \code{"package"} or \code{"help"}.
This selects the type of cache files to delete.}

\item{age_days}{A number that expresses a cache file's age that must
be exceeded for it to be deleted, with age defined as days elapsed since the
file's last modification date.
The default value of 90 days can also be changed with the option
\code{qgisprocess.cachefiles_days_keep} or the  environment variable
\code{R_QGISPROCESS_CACHEFILES_DAYS_KEEP}.}

\item{quiet}{Use \code{FALSE} to display more information,
possibly useful for debugging.}

\item{startup}{Logical.
Is this command being run while loading the package?}
}
\description{
Delete old cache files
}
\details{
Note that a currently used package cache file will never be deleted.

This function is called when loading the package.
}
\keyword{internal}
