% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-detect.R
\name{qgis_detect_windows}
\alias{qgis_detect_windows}
\alias{qgis_detect_macos}
\alias{is_macos}
\alias{is_windows}
\title{Detect QGIS installations with 'qgis_process'}
\usage{
qgis_detect_windows(drive_letter = strsplit(R.home(), ":")[[1]][1])

qgis_detect_macos()

is_macos()

is_windows()
}
\arguments{
\item{drive_letter}{The drive letter on which to search. By default,
this is the same drive letter as the R executable.}
}
\value{
A character vector of possible paths to the QGIS executable.
}
\description{
Detect QGIS installations with 'qgis_process'
}
\examples{
if (is_windows()) qgis_detect_windows()
if (is_macos()) qgis_detect_macos()

}
