% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-arguments.R
\name{qgis_make_arguments}
\alias{qgis_make_arguments}
\alias{qgis_serialize_arguments}
\alias{qgis_clean_arguments}
\alias{as_qgis_argument}
\alias{qgis_default_value}
\alias{is_qgis_default_value}
\alias{as_qgis_argument.default}
\alias{as_qgis_argument.qgis_default_value}
\alias{as_qgis_argument.NULL}
\alias{as_qgis_argument.character}
\alias{as_qgis_argument.logical}
\alias{as_qgis_argument.numeric}
\alias{qgis_clean_argument}
\alias{qgis_clean_argument.default}
\alias{qgis_clean_argument.qgis_tempfile_arg}
\alias{qgis_argument_spec_by_name}
\title{Type coercion for arguments to QGIS processing algorithms}
\usage{
qgis_make_arguments(algorithm, ...)

qgis_serialize_arguments(arguments)

qgis_clean_arguments(arguments)

as_qgis_argument(x, spec = qgis_argument_spec())

qgis_default_value()

is_qgis_default_value(x)

\method{as_qgis_argument}{default}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{qgis_default_value}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{`NULL`}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{character}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{logical}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{numeric}(x, spec = qgis_argument_spec())

qgis_clean_argument(value, spec = qgis_argument_spec())

\method{qgis_clean_argument}{default}(value, spec = qgis_argument_spec())

\method{qgis_clean_argument}{qgis_tempfile_arg}(value, spec = qgis_argument_spec())

qgis_argument_spec_by_name(
  algorithm,
  name,
  arguments = qgis_arguments(algorithm)
)
}
\arguments{
\item{algorithm}{A qualified algorithm name (e.g., "native:filedownloader") or
a path to a QGIS model file.}

\item{...}{Named key-value pairs as arguments for each algorithm. Features of
\code{\link[rlang:list2]{rlang::list2()}} are supported. These arguments
are converted to strings using \code{\link[=as_qgis_argument]{as_qgis_argument()}}.}

\item{arguments}{The result of \code{\link[=qgis_make_arguments]{qgis_make_arguments()}}.}

\item{x}{An object passed to a QGIS processing algorithm}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}

\item{value}{The result of \code{\link[=as_qgis_argument]{as_qgis_argument()}} after the QGIS processing
algorithm has been run.}
}
\description{
Calls to \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} can and should contain R objects that
need to be serialized before they are passed to the command line. In
some cases (e.g., sf objects), temporary files need to be written and
cleaned up. The \code{\link[=as_qgis_argument]{as_qgis_argument()}} and \code{\link[=qgis_clean_argument]{qgis_clean_argument()}} S3
generics provide hooks for argument values to be serialized correctly.
}
