% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-plugin.R
\name{qgis_plugins}
\alias{qgis_plugins}
\title{List, enable or disable QGIS plugins}
\usage{
qgis_plugins(which = "all", query = FALSE, quiet = TRUE)
}
\arguments{
\item{which}{String defining which plugins to enlist, based on their
status in QGIS (enabled or disabled).
Must be one of: \code{"all"}, \code{"enabled"}, \code{"disabled"}.}

\item{query}{Use \code{TRUE} to refresh the cached value.}

\item{quiet}{Use \code{FALSE} to display more information about the command, possibly
useful for debugging.}
}
\value{
A tibble of plugins and their status.
}
\description{
\code{qgis_plugins()} lists the available plugins that implement Processing
providers.
\code{qgis_enable_plugins()} enables plugins while \code{qgis_disable_plugins()}
does the reverse.
}
\details{
The cache is immediately updated upon enabling or disabling plugins from R.

If you installed, removed, enabled or disabled plugins in the GUI, then
run \code{qgis_configure()} to make those changes available in R.
}
\note{
Only plugins that implement Processing providers are supported.
Installing or removing plugins is not supported.
}
