% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-configure.R
\name{qgis_run}
\alias{qgis_run}
\alias{has_qgis}
\alias{qgis_configure}
\alias{qgis_unconfigure}
\alias{qgis_version}
\alias{qgis_path}
\alias{qgis_using_json_output}
\alias{qgis_using_json_input}
\title{Configure and run 'qgis_process'}
\usage{
qgis_run(args = character(), ..., env = qgis_env(), path = qgis_path())

has_qgis()

qgis_configure(quiet = FALSE, use_cached_data = FALSE)

qgis_unconfigure()

qgis_version(query = FALSE, quiet = TRUE, debug = FALSE)

qgis_path(query = FALSE, quiet = TRUE)

qgis_using_json_output(query = FALSE, quiet = TRUE)

qgis_using_json_input()
}
\arguments{
\item{args}{Command-line arguments}

\item{...}{Passed to \code{\link[processx:run]{processx::run()}}.}

\item{env}{A \code{\link[=list]{list()}} of environment variables.
Defaults to
\code{getOption("qgisprocess.env", list(QT_QPA_PLATFORM = "offscreen"))}.}

\item{path}{A path to the 'qgis_process' executable. Defaults to \code{\link[=qgis_path]{qgis_path()}}.}

\item{quiet}{Use \code{FALSE} to display more information about the command,
possibly useful for debugging.}

\item{use_cached_data}{Use the cached algorithm list and \code{path} found when
configuring qgisprocess during the last session. This saves some time
loading the package.}

\item{query}{Use \code{TRUE} to refresh the cached value.}

\item{debug}{Logical.
If \code{TRUE}, also output the version of QGIS, the operating system and all
relevant libraries, as reported by the 'qgis_process' command.}
}
\value{
The result of \code{\link[processx:run]{processx::run()}}.
}
\description{
Run \code{qgis_configure()} to bring the package configuration in line with
QGIS and to save this configuration to a persistent cache.
See the \emph{Details} section for more information about setting the path of
the 'qgis_process' command line tool.
\code{qgis_run()} is meant for directly calling this tool, but should normally not
be needed.
}
\details{
The qgisprocess package is a wapper around the 'qgis_process' command line
tool distributed with QGIS (>=3.14). Several functions use heuristics to
detect the location of the 'qgis_process' executable.

When loading the package, the configuration is automatically read from the
cache with \code{qgis_configure(use_cached_data = TRUE, quiet = TRUE)} in order
to save time.
Run \code{qgis_configure(use_cached_data = TRUE)} manually to get more details.

Use \code{qgis_algorithms()}, \code{qgis_providers()}, \code{qgis_plugins()},
\code{qgis_using_json_output()}, \code{qgis_path()} and \code{qgis_version()} to inspect cache
contents.

If the configuration
fails or you have more than one QGIS installation, you can set
\code{options(qgisprocess.path = "path/to/qgis_process")} or the
\code{R_QGISPROCESS_PATH} environment variable (useful on CI). On Linux the
'qgis_process' executable is generally available on the user's PATH,
on MacOS the executable is within the QGIS*.app/Contents/MacOS/bin folder,
and on Windows the executable is named qgis_process-qgis.bat or
qgis_process-qgis-dev.bat and is located in Program Files/QGIS*/bin or
OSGeo4W(64)/bin.
}
\examples{
if (has_qgis()) qgis_path()
if (has_qgis()) qgis_version()
if (has_qgis()) qgis_algorithms()
if (has_qgis()) qgis_providers()
if (has_qgis()) qgis_plugins()
qgis_configure(use_cached_data = TRUE)
qgis_configure()

}
