% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-arguments.R
\name{qgis_sanitize_arguments}
\alias{qgis_sanitize_arguments}
\alias{qgis_clean_arguments}
\alias{as_qgis_argument}
\alias{qgis_default_value}
\alias{is_qgis_default_value}
\alias{as_qgis_argument.default}
\alias{as_qgis_argument.list}
\alias{as_qgis_argument.qgis_default_value}
\alias{as_qgis_argument.NULL}
\alias{as_qgis_argument.character}
\alias{as_qgis_argument.logical}
\alias{as_qgis_argument.numeric}
\alias{qgis_clean_argument}
\alias{qgis_clean_argument.default}
\alias{qgis_clean_argument.qgis_tempfile_arg}
\alias{qgis_argument_spec_by_name}
\title{Type coercion for arguments to QGIS processing algorithms}
\usage{
qgis_sanitize_arguments(
  algorithm,
  ...,
  .algorithm_arguments = qgis_arguments(algorithm),
  .use_json_input = FALSE
)

qgis_clean_arguments(arguments)

as_qgis_argument(x, spec = qgis_argument_spec(), use_json_input = FALSE)

qgis_default_value()

is_qgis_default_value(x)

\method{as_qgis_argument}{default}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{list}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{qgis_default_value}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{`NULL`}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{character}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{logical}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

\method{as_qgis_argument}{numeric}(x, spec = qgis_argument_spec(), use_json_input = FALSE)

qgis_clean_argument(value)

\method{qgis_clean_argument}{default}(value)

\method{qgis_clean_argument}{qgis_tempfile_arg}(value)

qgis_argument_spec_by_name(
  algorithm,
  name,
  .algorithm_arguments = qgis_arguments(algorithm)
)
}
\arguments{
\item{.algorithm_arguments}{The result of \code{\link[=qgis_arguments]{qgis_arguments()}}}

\item{.use_json_input, use_json_input}{TRUE if the arguments will be
serialized as JSON instead of serialized as a command-line argument.}

\item{arguments}{The result of \code{\link[=qgis_sanitize_arguments]{qgis_sanitize_arguments()}}.}

\item{x}{An object passed to a QGIS processing algorithm}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}

\item{value}{The result of \code{\link[=as_qgis_argument]{as_qgis_argument()}} after the QGIS processing
algorithm has been run.}

\item{name}{The name of the input argument}
}
\description{
Calls to \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} can and should contain R objects that
need to be serialized before they are passed to the command line. In
some cases (e.g., sf objects), temporary files need to be written and
cleaned up. The \code{\link[=as_qgis_argument]{as_qgis_argument()}} and \code{\link[=qgis_clean_argument]{qgis_clean_argument()}} S3
generics provide hooks for argument values to be serialized correctly.
}
