% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-help.R
\name{qgis_show_help}
\alias{qgis_show_help}
\alias{qgis_get_description}
\alias{qgis_get_argument_specs}
\alias{qgis_get_output_specs}
\title{Get detailed information about one algorithm}
\usage{
qgis_show_help(algorithm)

qgis_get_description(algorithm)

qgis_get_argument_specs(algorithm)

qgis_get_output_specs(algorithm)
}
\arguments{
\item{algorithm}{A qualified algorithm name
(e.g., \code{"native:buffer"}).}
}
\value{
\itemize{
\item \code{qgis_get_description()}: a string.
\item \code{qgis_get_argument_specs()}, \code{qgis_get_output_specs()}: a tibble.
\item \code{qgis_show_help()}: the algorithm name, invisibly.
}
}
\description{
Get detailed information about one algorithm
}
\examples{
if (has_qgis()) {
  qgis_get_description("native:filedownloader")
  qgis_get_argument_specs("native:filedownloader")
  qgis_get_output_specs("native:filedownloader")
  qgis_show_help("native:filedownloader")
}

}
\seealso{
Other topics about information on algorithms & processing providers: 
\code{\link{qgis_algorithms}()},
\code{\link{qgis_search_algorithms}()}
}
\concept{topics about information on algorithms & processing providers}
