% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-process.R
\name{qgs_run}
\alias{qgs_run}
\alias{qgs_query_version}
\alias{qgs_query_algorithms}
\alias{qgs_path}
\alias{qgs_quiet}
\title{Configure and run 'qgis_process'}
\usage{
qgs_run(
  ...,
  timeout = 10,
  env = list(QT_QPA_PLATFORM = "offscreen"),
  quiet = qgs_quiet()
)

qgs_query_version(timeout = 10)

qgs_query_algorithms(timeout = 10)

qgs_path()

qgs_quiet()
}
\arguments{
\item{..., timeout}{Passed to \code{\link[processx:run]{processx::run()}}.}

\item{env}{A \code{\link[=list]{list()}} of environment variables.}

\item{quiet}{Use \code{FALSE} to display more information about the command, possibly
useful for debugging. Defaults to the value of \code{\link[=qgs_quiet]{qgs_quiet()}}.}
}
\value{
The result of \code{\link[processx:run]{processx::run()}}.
}
\description{
Configure and run 'qgis_process'
}
\examples{

}
