% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-state.R
\name{resolve_explicit_json_output}
\alias{resolve_explicit_json_output}
\title{Handle an explicitly set 'use_json_output'}
\usage{
resolve_explicit_json_output(json_output_setting, qgis_version)
}
\description{
The \code{qgisprocess.use_json_output} option or the
\code{R_QGISPROCESS_USE_JSON_OUTPUT} environment variable
can be set explicitly by the user. The function determines whether the user
setting can be accepted, depending on the presence of an explicit user
setting of the \code{qgisprocess.use_json_input} option or the
\code{R_QGISPROCESS_USE_JSON_INPUT} environment variable.
}
\details{
This helper already assumes that \code{qgisprocess.use_json_output} or
\code{R_QGISPROCESS_USE_JSON_OUTPUT} have an
explicit setting; the empty case is not being handled here!

This function intercepts the potential conflict of json_input being
explicitly set as \code{TRUE} and json_output being explicitly set as \code{FALSE}.
In such case, the result will still be set as \code{TRUE} on condition that
the json_input setting is valid (see \code{\link[=qgis_using_json_input]{qgis_using_json_input()}}).
}
\keyword{internal}
