% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-stars.R
\name{st_as_stars}
\alias{st_as_stars}
\alias{st_as_stars.qgis_outputRaster}
\alias{st_as_stars.qgis_outputLayer}
\alias{st_as_stars.qgis_result}
\title{Convert a qgis_result object or one of its elements to a stars object}
\usage{
st_as_stars.qgis_outputRaster(.x, ...)

st_as_stars.qgis_outputLayer(.x, ...)

st_as_stars.qgis_result(.x, ...)
}
\arguments{
\item{.x}{A result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} or one of the
\code{\link[=qgis_extract_output]{qgis_extract_output()}} functions.}

\item{...}{Arguments passed to \code{\link[stars:read_stars]{stars::read_stars()}}.}
}
\description{
Convert a qgis_result object or one of its elements to a stars object
}
\note{
Just use \code{st_as_stars()} in R scripts, it will use the correct
method.
}
\seealso{
Other topics about coercing processing output: 
\code{\link{qgis_as_raster}()},
\code{\link{qgis_as_terra}()},
\code{\link{st_as_sf}}

Other topics about accessing or managing processing results: 
\code{\link{qgis_as_raster}()},
\code{\link{qgis_as_terra}()},
\code{\link{qgis_clean_result}()},
\code{\link{qgis_extract_output}()},
\code{\link{qgis_result_status}()},
\code{\link{st_as_sf}}
}
\concept{topics about accessing or managing processing results}
\concept{topics about coercing processing output}
