% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-raster.R, R/qgis-arguments.R
\name{as_qgis_argument.RasterLayer}
\alias{as_qgis_argument.RasterLayer}
\alias{as_qgis_argument.RasterBrick}
\alias{as_qgis_argument}
\alias{qgis_default_value}
\alias{is_qgis_default_value}
\alias{as_qgis_argument.default}
\alias{as_qgis_argument.qgis_default_value}
\alias{as_qgis_argument.NULL}
\alias{as_qgis_argument.character}
\alias{as_qgis_argument.logical}
\alias{as_qgis_argument.numeric}
\alias{qgis_clean_argument}
\alias{qgis_clean_argument.default}
\alias{qgis_clean_argument.qgis_tempfile_arg}
\alias{qgis_argument_spec}
\alias{qgis_argument_spec_by_name}
\title{Type coercion for arguments to QGIS processing algorithms}
\usage{
\method{as_qgis_argument}{RasterLayer}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{RasterBrick}(x, spec = qgis_argument_spec())

as_qgis_argument(x, spec = qgis_argument_spec())

qgis_default_value()

is_qgis_default_value(x)

\method{as_qgis_argument}{default}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{qgis_default_value}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{`NULL`}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{character}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{logical}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{numeric}(x, spec = qgis_argument_spec())

qgis_clean_argument(value, spec = qgis_argument_spec())

\method{qgis_clean_argument}{default}(value, spec = qgis_argument_spec())

\method{qgis_clean_argument}{qgis_tempfile_arg}(value, spec = qgis_argument_spec())

qgis_argument_spec(
  algorithm = NA_character_,
  name = NA_character_,
  description = NA_character_,
  qgis_type = NA_character_
)

qgis_argument_spec_by_name(
  algorithm,
  name,
  arguments = qgis_arguments(algorithm)
)
}
\arguments{
\item{x}{An object passed to a QGIS processing algorithm}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}

\item{value}{The result of \code{\link[=as_qgis_argument]{as_qgis_argument()}} after the QGIS processing
algorithm has been run.}

\item{algorithm}{A qualified algorithm name (e.g., "native:filedownloader") or
a path to a QGIS model file.}

\item{name, description, qgis_type}{Column values of \code{arguments} denoting
the argument name, its description, and its QGIS type name.}

\item{arguments}{The result of \code{\link[=qgis_arguments]{qgis_arguments()}}.}
}
\description{
Calls to \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} can and should contain R objects that
need to be serialized before they are passed to the command line. In
some cases (e.g., sf objects), temporary files need to be written and
cleaned up. The \code{\link[=as_qgis_argument]{as_qgis_argument()}} and \code{\link[=qgis_clean_argument]{qgis_clean_argument()}} S3
generics provide hooks for argument values to be serialized correctly.
}
