% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-stars.R
\name{as_qgis_argument.stars}
\alias{as_qgis_argument.stars}
\alias{as_qgis_argument.stars_proxy}
\title{Convert raster objects to/from QGIS inputs/outputs}
\usage{
\method{as_qgis_argument}{stars}(x, spec = qgis_argument_spec())

\method{as_qgis_argument}{stars_proxy}(x, spec = qgis_argument_spec())
}
\arguments{
\item{x}{A stars or stars_proxy object.}

\item{spec}{A \code{list()} with values for \code{algorithm}, \code{name},
\code{description}, and \code{qgis_type}. See \code{\link[=qgis_argument_spec]{qgis_argument_spec()}} to
create a blank \code{spec} for testing.}
}
\description{
Convert raster objects to/from QGIS inputs/outputs
}
