% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-result.R
\name{is_qgis_result}
\alias{is_qgis_result}
\alias{qgis_result_clean}
\alias{qgis_extract_output_by_name}
\alias{qgis_extract_output_by_position}
\alias{qgis_extract_output_by_class}
\alias{qgis_result_status}
\alias{qgis_result_stdout}
\alias{qgis_result_stderr}
\alias{qgis_result_args}
\title{Access algorithm results}
\usage{
is_qgis_result(x)

qgis_result_clean(x)

qgis_extract_output_by_name(x, name = "OUTPUT", first = TRUE)

qgis_extract_output_by_position(x, which)

qgis_extract_output_by_class(x, class, single = TRUE)

qgis_result_status(x)

qgis_result_stdout(x)

qgis_result_stderr(x)

qgis_result_args(x)
}
\arguments{
\item{x}{An object returned by \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}.}

\item{name}{The name of an output.}

\item{first}{Logical.
Should \code{qgis_extract_output_by_name()} fall back to the first
output element if the default \code{OUTPUT} or \code{output} element are not available?
Only takes effect if \code{name} is not found.}

\item{which}{The index of an output.}

\item{class}{Character vector of classes.
At least one class must be inherited by an element of \code{x} for that element
to be selected.}

\item{single}{Logical.
Ensures the selection of a single output in \code{qgis_extract_output_by_class()}.
The \code{OUTPUT} or \code{output} element is taken if available; otherwise falls back
to the first element (that meets the specified class).}
}
\description{
Access algorithm results
}
