% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-raster.R
\name{qgis_as_raster}
\alias{qgis_as_raster}
\alias{qgis_as_brick}
\alias{qgis_as_raster.qgis_outputRaster}
\alias{qgis_as_brick.qgis_outputRaster}
\alias{qgis_as_raster.qgis_outputLayer}
\alias{qgis_as_brick.qgis_outputLayer}
\alias{qgis_as_raster.qgis_result}
\alias{qgis_as_brick.qgis_result}
\title{Convert a qgis_result object or one of its elements to a raster object}
\usage{
qgis_as_raster(output, ...)

qgis_as_brick(output, ...)

\method{qgis_as_raster}{qgis_outputRaster}(output, ...)

\method{qgis_as_brick}{qgis_outputRaster}(output, ...)

\method{qgis_as_raster}{qgis_outputLayer}(output, ...)

\method{qgis_as_brick}{qgis_outputLayer}(output, ...)

\method{qgis_as_raster}{qgis_result}(output, ...)

\method{qgis_as_brick}{qgis_result}(output, ...)
}
\arguments{
\item{output}{A result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} or one of the
\code{\link[=qgis_extract_output]{qgis_extract_output()}} functions.}

\item{...}{Arguments passed to \code{\link[raster:raster]{raster::raster()}} or \code{\link[raster:brick]{raster::brick()}}.}
}
\description{
Convert a qgis_result object or one of its elements to a raster object
}
\note{
Just use \code{qgis_as_raster()} and \code{qgis_as_brick()} in R scripts;
it will use the correct method.
}
\seealso{
Other topics about coercing processing output: 
\code{\link{qgis_as_terra}()},
\code{\link{st_as_sf}},
\code{\link{st_as_stars}}

Other topics about accessing or managing processing results: 
\code{\link{qgis_as_terra}()},
\code{\link{qgis_clean_result}()},
\code{\link{qgis_extract_output}()},
\code{\link{qgis_result_status}()},
\code{\link{st_as_sf}},
\code{\link{st_as_stars}}
}
\concept{topics about accessing or managing processing results}
\concept{topics about coercing processing output}
