% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-function.R
\name{qgis_function}
\alias{qgis_function}
\title{Create functions from QGIS algorithms}
\usage{
qgis_function(algorithm, ...)
}
\arguments{
\item{algorithm}{A qualified algorithm name
(e.g., "native:filedownloader").}

\item{...}{Algorithm arguments.
These values are evaluated once and immediately, so you shouldn't
call \code{\link[=qgis_tmp_file]{qgis_tmp_file()}} here.}
}
\description{
As opposed to \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}, \code{\link[=qgis_function]{qgis_function()}} creates a callable
function based on the argument metadata provided by \code{\link[=qgis_get_argument_specs]{qgis_get_argument_specs()}}.
Unlike \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}, \code{\link[=qgis_function]{qgis_function()}} sets the default value
of \code{.quiet} to \code{TRUE} to make the function more usable within other
R code.
}
\examples{
if (has_qgis()) {
  qgis_buffer <- qgis_function("native:buffer")
  qgis_buffer(
    system.file(
      "longlake/longlake_depth.gpkg",
      package = "qgisprocess"
    ),
    DISTANCE = 10
  )
}

}
