% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-algorithm.R
\name{qgis_has_algorithm}
\alias{qgis_has_algorithm}
\alias{qgis_has_provider}
\alias{qgis_providers}
\alias{is_qgis_model_file}
\alias{assert_qgis_algorithm_or_model_file}
\title{Run algorithms using 'qgis_process'}
\usage{
qgis_has_algorithm(algorithm)

qgis_has_provider(provider)

qgis_providers(provider)

is_qgis_model_file(algorithm)

assert_qgis_algorithm_or_model_file(algorithm)
}
\arguments{
\item{algorithm}{A qualified algorithm name (e.g., "native:filedownloader") or
a path to a QGIS model file.}

\item{provider}{A provider identifier (e.g., "native")}
}
\description{
Run algorithms using 'qgis_process'
}
\examples{
if (has_qgis()) qgis_has_algorithm("native:filedownloader")
if (has_qgis()) qgis_has_provider("native")
if (has_qgis()) qgis_providers()

}
