% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-arguments.R
\name{qgis_list_input}
\alias{qgis_list_input}
\alias{qgis_dict_input}
\title{Specify lists and dictionary inputs}
\usage{
qgis_list_input(...)

qgis_dict_input(...)
}
\arguments{
\item{...}{Named values for \code{\link[=qgis_dict_input]{qgis_dict_input()}} or
unnamed values for \code{\link[=qgis_list_input]{qgis_list_input()}}.}
}
\value{
\itemize{
\item \code{qgis_list_input()}: An object of class 'qgis_list_input'
\item \code{qgis_dict_input()}: An object of class 'qgis_dict_input'
}
}
\description{
Specify lists and dictionary inputs
}
\examples{
qgis_list_input(1, 2, 3)
qgis_dict_input(a = 1, b = 2, c = 3)

}
