% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-configure.R
\name{qgis_run}
\alias{qgis_run}
\alias{has_qgis}
\alias{assert_qgis}
\alias{qgis_configure}
\alias{qgis_unconfigure}
\alias{qgis_version}
\alias{qgis_path}
\alias{qgis_query_path}
\alias{qgis_env}
\alias{qgis_query_version}
\alias{qgis_query_algorithms}
\title{Configure and run 'qgis_process'}
\usage{
qgis_run(args = character(), ..., env = qgis_env(), path = qgis_path())

has_qgis()

assert_qgis(action = abort)

qgis_configure(quiet = FALSE)

qgis_unconfigure()

qgis_version(query = FALSE, quiet = TRUE)

qgis_path(query = FALSE, quiet = TRUE)

qgis_query_path(quiet = FALSE)

qgis_env()

qgis_query_version(quiet = FALSE)

qgis_query_algorithms(quiet = FALSE)
}
\arguments{
\item{args}{Command-line arguments}

\item{...}{Passed to \code{\link[processx:run]{processx::run()}}.}

\item{env}{A \code{\link[=list]{list()}} of environment variables. Defaults to \code{\link[=qgis_env]{qgis_env()}}.}

\item{path}{A path to the 'qgis_process' executable. Defaults to \code{\link[=qgis_path]{qgis_path()}}.}

\item{action}{An action to take if the 'qgis_process' executable could not be
found.}

\item{quiet}{Use \code{FALSE} to display more information about the command, possibly
useful for debugging.}

\item{query}{Use \code{TRUE} to refresh the cached value.}
}
\value{
The result of \code{\link[processx:run]{processx::run()}}.
}
\description{
The qgisprocess package is a wapper around the 'qgis_process' command line
tool distributed with QGIS (>=3.14). These functions use heuristics to
detect the location of the 'qgis_process' executable. If the configuration
fails or you have more than one QGIS installation, you can set
\code{options(qgisprocess.path = "path/to/qgis_process")} or the
\code{R_QGISPROCESS_PATH} environment variable (useful on CI). On Linux the
'qgis_process' executable is generally available on the user's PATH,
on MacOS the executable is within the QGIS*.app/Contents/MacOS/bin folder,
and on Windows the executable is named qgis_process-qgis.bat or
qgis_process-qgis-dev.bat and is located in Program Files/QGIS*/bin or
OSGeo4W(64)/bin.
}
\examples{
if (has_qgis()) qgis_path()
if (has_qgis()) qgis_version()
if (has_qgis()) qgis_algorithms()
qgis_configure()

}
