% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-configure.R
\name{qgis_run}
\alias{qgis_run}
\alias{has_qgis}
\alias{assert_qgis}
\alias{qgis_configure}
\alias{qgis_version}
\alias{qgis_algorithms}
\alias{qgis_path}
\alias{qgis_query_path}
\alias{qgis_env}
\alias{qgis_query_version}
\alias{qgis_query_algorithms}
\title{Configure and run 'qgis_process'}
\usage{
qgis_run(..., env = qgis_env(), path = qgis_path())

has_qgis()

assert_qgis(action = stop)

qgis_configure(quiet = FALSE)

qgis_version(query = FALSE, quiet = TRUE)

qgis_algorithms(query = FALSE, quiet = TRUE)

qgis_path(query = FALSE, quiet = TRUE)

qgis_query_path(quiet = FALSE)

qgis_env()

qgis_query_version(quiet = FALSE)

qgis_query_algorithms(quiet = FALSE)
}
\arguments{
\item{...}{Passed to \code{\link[processx:run]{processx::run()}}.}

\item{env}{A \code{\link[=list]{list()}} of environment variables. Defaults to \code{\link[=qgis_env]{qgis_env()}}.}

\item{path}{A path to the 'qgis_process' executable. Defaults to \code{\link[=qgis_path]{qgis_path()}}.}

\item{action}{An action to take if the 'qgis_process' executable could not be
found.}

\item{quiet}{Use \code{FALSE} to display more information about the command, possibly
useful for debugging.}

\item{query}{Use \code{TRUE} to refresh the cached value.}
}
\value{
The result of \code{\link[processx:run]{processx::run()}}.
}
\description{
Configure and run 'qgis_process'
}
\examples{
if (has_qgis()) qgis_path()
if (has_qgis()) qgis_version()
if (has_qgis()) qgis_algorithms()

}
