% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-help.R
\name{qgis_show_help}
\alias{qgis_show_help}
\alias{qgis_description}
\alias{qgis_arguments}
\alias{qgis_outputs}
\alias{qgis_help}
\title{Show algorithm help}
\usage{
qgis_show_help(algorithm)

qgis_description(algorithm)

qgis_arguments(algorithm)

qgis_outputs(algorithm)

qgis_help(algorithm)
}
\description{
Show algorithm help
}
\examples{
if (has_qgis()) qgis_show_help("native:filedownloader")
if (has_qgis()) qgis_description("native:filedownloader")
if (has_qgis()) qgis_arguments("native:filedownloader")

}
