% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-help.R
\name{qgis_show_help}
\alias{qgis_show_help}
\alias{qgis_get_description}
\alias{qgis_get_argument_specs}
\alias{qgis_get_output_specs}
\title{Show algorithm help}
\usage{
qgis_show_help(algorithm)

qgis_get_description(algorithm)

qgis_get_argument_specs(algorithm)

qgis_get_output_specs(algorithm)
}
\arguments{
\item{algorithm}{A qualified algorithm name (e.g., "native:filedownloader") or
a path to a QGIS model file.}
}
\description{
Show algorithm help
}
\examples{
if (has_qgis()) qgis_show_help("native:filedownloader")
if (has_qgis()) qgis_get_description("native:filedownloader")
if (has_qgis()) qgis_get_argument_specs("native:filedownloader")

}
