% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-help.R
\name{qgis_show_help}
\alias{qgis_show_help}
\alias{qgis_get_description}
\alias{qgis_get_argument_specs}
\alias{qgis_get_output_specs}
\title{Get detailed information about one algorithm}
\usage{
qgis_show_help(algorithm)

qgis_get_description(algorithm)

qgis_get_argument_specs(algorithm)

qgis_get_output_specs(algorithm)
}
\arguments{
\item{algorithm}{A qualified algorithm name
(e.g., \code{"native:buffer"}).}
}
\description{
Get detailed information about one algorithm
}
\examples{
if (has_qgis()) qgis_show_help("native:filedownloader")
if (has_qgis()) qgis_get_description("native:filedownloader")
if (has_qgis()) qgis_get_argument_specs("native:filedownloader")

}
\seealso{
Other topics about information on algorithms & processing providers: 
\code{\link{qgis_algorithms}()}
}
\concept{topics about information on algorithms & processing providers}
