% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis-configure.R
\name{qgis_unconfigure}
\alias{qgis_unconfigure}
\title{Clean the package cache}
\usage{
qgis_unconfigure()
}
\value{
\code{NULL}, invisibly.
}
\description{
Empties the qgisprocess cache environment.
}
\examples{
\dontrun{
# not running this function in example() as it clears the cache environment.
qgis_unconfigure()
}

# undoing qgis_unconfigure() by repopulating the cache environment from file:
\donttest{
# not running in R CMD check to save time
qgis_configure(use_cached_data = TRUE)
}

}
\seealso{
Other topics about programming or debugging utilities: 
\code{\link{qgis_result_status}()},
\code{\link{qgis_run}()},
\code{\link{qgis_tmp_file}()},
\code{\link{qgis_using_json_input}()}
}
\concept{topics about programming or debugging utilities}
