% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgisprocess-deprecated.R
\name{qgis_output}
\alias{qgis_output}
\title{Deprecated functions}
\usage{
qgis_output(x, which)
}
\arguments{
\item{x}{An object returned by \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}.}

\item{which}{The index of an output.}
}
\description{
Deprecated functions
}
\keyword{internal}
