% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgisprocess-package.R
\docType{package}
\name{qgisprocess-package}
\alias{qgisprocess}
\alias{qgisprocess-package}
\title{qgisprocess: Use 'QGIS' Processing Algorithms}
\description{
Provides seamless access to the 'QGIS' \url{https://qgis.org/} processing toolbox using the 'qgis_process' command-line utility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatial/qgisprocess}
  \item Report bugs at \url{https://github.com/r-spatial/qgisprocess/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (\href{https://orcid.org/0000-0002-6378-6229}{ORCID})
  \item Jan Caha \email{jan.caha@outlook.com} (\href{https://orcid.org/0000-0003-0165-0606}{ORCID})
}

Other contributors:
\itemize{
  \item Antony Barja \email{antony.barja8@gmail.com} (\href{https://orcid.org/0000-0001-5921-2858}{ORCID}) [contributor]
  \item Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID}) [contributor]
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID}) [contributor]
}

}
\keyword{internal}
