% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgisprocess-package.R
\docType{package}
\name{qgisprocess-package}
\alias{qgisprocess}
\alias{qgisprocess-package}
\title{qgisprocess: Use 'QGIS' Processing Algorithms}
\description{
Provides seamless access to the 'QGIS' (\url{https://qgis.org/en/site/}) processing toolbox using the standalone 'qgis_process' command-line utility. Both native and third-party (plugin) processing providers are supported. Beside referring data sources from file, also common objects from 'sf', 'terra' and 'stars' are supported. The native processing algorithms are documented at \url{https://docs.qgis.org/latest/en/docs/user_manual/processing_algs/}.
}
\note{
A structured overview of the available functions can be consulted at
\url{https://r-spatial.github.io/qgisprocess/reference/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-spatial.github.io/qgisprocess/}
  \item \url{https://github.com/r-spatial/qgisprocess}
  \item Report bugs at \url{https://github.com/r-spatial/qgisprocess/issues}
}

}
\author{
\strong{Maintainer}: Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (\href{https://orcid.org/0000-0002-6378-6229}{ORCID}) (Research Institute for Nature and Forest (INBO))

Authors:
\itemize{
  \item Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID}) (Voltron Data)
  \item Jan Caha \email{jan.caha@outlook.com} (\href{https://orcid.org/0000-0003-0165-0606}{ORCID})
  \item Jannes Muenchow \email{malnamalja@gmx.de} (\href{https://orcid.org/0000-0001-7834-4717}{ORCID})
}

Other contributors:
\itemize{
  \item Antony Barja \email{antony.barja8@gmail.com} (\href{https://orcid.org/0000-0001-5921-2858}{ORCID}) [contributor]
  \item Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID}) [contributor]
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID}) [contributor]
}

}
\keyword{internal}
