% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgisprocess-package.R
\docType{package}
\name{qgisprocess-package}
\alias{qgisprocess}
\alias{qgisprocess-package}
\title{qgisprocess: Use 'QGIS' Processing Algorithms}
\description{
Provides seamless access to the 'QGIS' <https://qgis.org/>
  processing toolbox using the 'qgis_process' command-line utility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/qgisprocess}
  \item Report bugs at \url{https://github.com/paleolimbot/qgisprocess/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Jan Caha
}

Other contributors:
\itemize{
  \item Robin Lovelace (\href{https://orcid.org/0000-0001-5679-6536}{ORCID}) [contributor]
}

}
\keyword{internal}
