% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-sf.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.qgis_result}
\title{Convert a qgis_result object to an sf object}
\usage{
st_as_sf.qgis_result(x, ...)
}
\arguments{
\item{x}{A result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}}.}

\item{...}{Arguments passed to \code{\link[sf:st_read]{sf::read_sf()}}.}
}
\description{
Convert a qgis_result object to an sf object
}
