% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-stars.R
\name{st_as_stars}
\alias{st_as_stars}
\alias{st_as_stars.qgis_outputRaster}
\alias{st_as_stars.qgis_outputLayer}
\alias{st_as_stars.qgis_result}
\title{Convert a qgis_result object or one of its elements to a stars object}
\usage{
st_as_stars.qgis_outputRaster(.x, ...)

st_as_stars.qgis_outputLayer(.x, ...)

st_as_stars.qgis_result(.x, ...)
}
\arguments{
\item{.x}{A result from \code{\link[=qgis_run_algorithm]{qgis_run_algorithm()}} or one of the
\code{\link[=qgis_extract_output]{qgis_extract_output()}} functions.}

\item{...}{Arguments passed to \code{\link[stars:read_stars]{stars::read_stars()}}.}
}
\description{
Convert a qgis_result object or one of its elements to a stars object
}
